requirements.txt
# YouTube Maker - Required Packages

# GUI Framework
ttkbootstrap==1.10.1

# Image Processing
Pillow>=10.0.0

# HTTP Requests
requests>=2.31.0

# YouTube API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Gemini API
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0

# EXE Building (optional, for distribution)
pyinstaller>=6.0.0